Command 'scoop update' delete scoop if download fail.  (#1120)
* try catch on update

fixes #1118

* deleting the right dir

* added -ea stop to git command

* added try catch to bucket update

* removing space

* Removing -ea from git command, since it isn´t a scriptlet

* When git clone fail, it does not create de $newdir, so I changed the code
to check for this path.

When the git pull fail, nothing happens, so It doesnt need any changes

Also, git error could not be catch in powersheel exception

* removing blank lines

+++ libexec/scoop-update.ps1
@@ -59,6 +59,11 @@ function update_scoop() {
         # get git scoop
         git_clone -q $repo --branch $branch --single-branch "`"$newdir`""
 
+        # check if scoop was successful downloaded
+        if(!(test-path "$newdir")) {
+            abort 'scoop update failed'
+        }
+
         # replace non-git scoop with the git version
         rm -r -force $currentdir -ea stop
         mv $newdir $currentdir
